server/app/tests/auth/test_logout_view.py
from django.urls import reverse
from rest_framework import status
from rest_framework.test import APITestCase
from rest_framework_simplejwt.tokens import RefreshToken
from django.contrib.auth.models import User


class LogoutViewTests(APITestCase):
    def setUp(self):
        self.user = User.objects.create_user(username='testuser', password='testpassword')
        self.token = RefreshToken.for_user(self.user)

    def test_logout_user(self):
        url = reverse('logout')
        data = {'refresh': str(self.token)}
        self.client.force_authenticate(user=self.user)
        response = self.client.post(url, data, format='json')
        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
        self.assertFalse(RefreshToken(self.token).blacklisted)

    def test_logout_user_unauthenticated(self):
        url = reverse('logout')
        data = {'refresh': str(self.token)}
        response = self.client.post(url, data, format='json')
        self.assertEqual(response.status_code, status.HTTP_401_UNAUTHORIZED)
        self.assertTrue(RefreshToken(self.token).blacklisted)