server/requirements.txt
arrow==1.2.3
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.2.0
colorama==0.4.6
daphne==4.1.2
Django==4.2.2
django-auditlog==3.0.0
django-cors-headers==4.3.1
django-channels-jwt-auth-middleware==1.0.0
django-filter==24.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.4.2
drf-extensions==0.7.1
drf-spectacular==0.26.5
Faker==24.4.0
idna==3.4
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
loguru==0.7.1
pillow==10.2.0
psycopg2-binary==2.9.9
pydash==7.0.4
PyJWT==2.8.0
pymongo==4.4.1
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
PyYAML==6.0.1
redis==5.0.4
referencing==0.31.0
requests==2.31.0
rpds-py==0.13.1
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
win32-setctime==1.1.0
pandas==2.2.2
openpyxl==3.1.2